# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Animal Adoption API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the animal adoption application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Adopter": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "phone"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the adopter",
            "example": "adopter_123abc"
          },
          "name": {
            "type": "string",
            "description": "Full name of the adopter",
            "example": "Jane Smith"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the adopter",
            "example": "jane@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Contact number of the adopter",
            "example": "+1-204-555-1234"
          },
          "address": {
            "type": "string",
            "description": "Home address of the adopter",
            "example": "123 Main St, Winnipeg, MB"
          },
          "preferredSpecies": {
            "type": "array",
            "description": "Preferred animal species for adoption",
            "items": {
              "type": "string"
            },
            "example": [
              "Dog",
              "Cat"
            ]
          }
        }
      },
      "CreateAdopterRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phone"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the adopter",
            "example": "Jane Smith"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the adopter",
            "example": "jane@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Contact number of the adopter",
            "example": "+1-204-555-1234"
          },
          "address": {
            "type": "string",
            "description": "Home address of the adopter",
            "example": "123 Main St, Winnipeg, MB"
          },
          "preferredSpecies": {
            "type": "array",
            "description": "Preferred animal species for adoption",
            "items": {
              "type": "string"
            },
            "example": [
              "Dog",
              "Cat"
            ]
          }
        }
      },
      "UpdateAdopterRequest": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "preferredSpecies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/adopters": {
      "post": {
        "summary": "Create a new adopter",
        "tags": [
          "Adopters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Only admin and staff users can create adopters",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdopterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Adopter created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Adopter created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Adopter"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Invalid adopter data"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "get": {
        "summary": "Retrieve all adopters",
        "tags": [
          "Adopters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Adopters retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Adopters retrieved successfully"
                    },
                    "count": {
                      "type": "integer",
                      "example": 5
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Adopter"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/adopters/{id}": {
      "get": {
        "summary": "Retrieve an adopter by ID",
        "tags": [
          "Adopters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Adopter ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Adopter retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Adopter retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Adopter"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Adopter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an adopter",
        "tags": [
          "Adopters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Only admins can update adopters",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Adopter ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdopterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Adopter updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Adopter updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Adopter"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Adopter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an adopter",
        "tags": [
          "Adopters"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Only admins can delete adopters",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Adopter ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Adopter deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Adopter deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Adopter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}